把函数都放到utils.ts里，避免循环依赖
